requirements.txt
fastapi
uvicorn
python-multipart
paddlepaddle
paddleocr
google-generativeai
pillow
pydantic
pydantic-settings
python-dotenv
requests
sqlalchemy
